IntroPage 디자인 변경, Container 컴포넌트 분리
