+++ src/HomePage.jsx
@@ -10,38 +10,7 @@ import { keyframes } from '@emotion/react'
 
 import { illustrations } from './assets/images';
 
-const Container = styled.div({
-  display: 'grid',
-  height: '80%',
-  gridTemplateColumns: 'repeat(2, 1fr)',
-  gridTemplateRows: '1fr',
-});
-
-const BottleImg = styled.img({
-  width: '90%',
-  height: '80%',
-  margin: 'auto',
-});
-
-const Wrapper = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  textAlign: 'center',
-  justifyContent: 'space-around',
-  margin: '15% 0',
-});
-
-const Title = styled.h2({
-  color: '#FFB61D',
-  fontSize: '70px',
-  margin: '0',
-});
-
-const Text = styled.p({
-  color: '#323232',
-  fontSize: '25px',
-  lineHeight: '50px',
-});
+import Container from './styles/Container';
 
 const heartBeat = keyframes`
   0%{
@@ -77,17 +46,17 @@ const Start = styled.div({
 export default function HomePage() {
   return (
     <Container>
-      <BottleImg src={illustrations[1]} />
-      <Wrapper>
-        <Title>Vitamin Y</Title>
-        <Text>
+      <img src={illustrations[1]} />
+      <div>
+        <h2>Vitamin Y</h2>
+        <p className="homeText">
           당신의 비타민 Y를 찾아보세요
           <br />
           이름을 알려주실래요?
-        </Text>
+        </p>
         <HomeContainer />
         <Start><Link to="/intro"><i className="fas fa-tablets"></i></Link></Start>
-      </Wrapper>
+      </div>
     </Container>
   );
 }
